==> VGR-COVID-19/1.Mortality-Cohort.ipynb <==
import pandas as pd

C19_ID = pd.read_csv('data/COVID_19_ID_full.csv')
c_list = C19_ID.Patientgrupp.unique().tolist() 
display(c_list)
covid_test_results = C19_ID[C19_ID.Patientgrupp==c_list[0]].Observation.value_counts().index.values[0:9].tolist()
print('Test results: \n', covid_test_results)
covid_pos_results = covid_test_results[0:4] + covid_test_results[5:7] 
print('Keys for pos result: \n', covid_pos_results)

# Select patients and info
c_patients = C19_ID[(C19_ID.Patientgrupp == c_list[0]) &
                    (C19_ID.Observation.isin(covid_pos_results))
                   ]
print("Number of patients: ", c_patients.shape[0])

# Filter some patients/data

# Drop underage patients
c_patients = c_patients[c_patients.Ålder > 18]

# Keep only some columns
c_patients = c_patients[['Patient_ID','Ålder','Första_kontaktdatum']]
c_patients.drop_duplicates(subset=['Patient_ID'], inplace = True)
c_patients.reset_index(inplace=True, drop=True)

print(c_patients.shape[0])

print(c_patients['Första_kontaktdatum'].max())
print(c_patients['Första_kontaktdatum'].min())

c_patients.to_csv('./data_gen/cohort.csv', index=False)
